Modify: 모임일 투표 패키지 변경
voting -> date 패키지로 변경
date 패키지는 2개의 도메인을 가짐.
voting 도메인과 confirm 도메인
voting 도메인은 모임일 투표 관련.
confirm 도메인은 모임일 확정 관련.
