+++ src/main/java/com/comeon/backend/date/api/dto/RemoveVotingResponse.java
@@ -1,0 +1,15 @@
+package com.comeon.backend.date.api.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class RemoveVotingResponse {
+
+    private Boolean success;
+
+    public RemoveVotingResponse() {
+        this.success = true;
+    }
+}
