+++ src/main/java/com/comeon/backend/date/command/domain/voting/MeetingCalendarService.java
@@ -1,0 +1,8 @@
+package com.comeon.backend.date.command.domain.voting;
+
+import java.time.LocalDate;
+
+public interface MeetingCalendarService {
+
+    boolean verifyDateInMeetingCalendar(Long meetingId, LocalDate date);
+}
