+++ src/main/java/com/comeon/backend/date/query/dto/VotingSimpleResponse.java
@@ -1,0 +1,20 @@
+package com.comeon.backend.date.query.dto;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDate;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class VotingSimpleResponse {
+
+    @JsonFormat(pattern = "yyyy-MM-dd", timezone = "Asia/Seoul")
+    private LocalDate date;
+
+    private int memberCount;
+    private boolean myVoting;
+}
