--- src/main/java/com/comeon/backend/voting/command/application/DateOutOfMeetingCalendarException.java
@@ -1,18 +1,0 @@
-package com.comeon.backend.voting.command.application;
-
-import com.comeon.backend.common.error.ErrorCode;
-import com.comeon.backend.common.error.MeetingErrorCode;
-import com.comeon.backend.common.error.RestApiException;
-
-public class DateOutOfMeetingCalendarException extends RestApiException {
-
-    private static final ErrorCode errorCode = MeetingErrorCode.DATE_OUT_OF_CALENDAR_RANGE;
-
-    public DateOutOfMeetingCalendarException() {
-        super(errorCode);
-    }
-
-    public DateOutOfMeetingCalendarException(String message) {
-        super(message, errorCode);
-    }
-}
