--- src/main/java/com/comeon/backend/voting/command/domain/DateVotingRepository.java
@@ -1,12 +1,0 @@
-package com.comeon.backend.voting.command.domain;
-
-import java.time.LocalDate;
-import java.util.Optional;
-
-public interface DateVotingRepository {
-
-    DateVoting save(DateVoting dateVoting);
-    Optional<DateVoting> findDateVotingBy(Long meetingId, Long userId, LocalDate date);
-    void deleteDateVotingBy(Long meetingId, Long userId, LocalDate date);
-    void deleteDateVotingBy(DateVoting dateVoting);
-}
