--- src/main/java/com/comeon/backend/voting/command/domain/MeetingCalendarService.java
@@ -1,8 +1,0 @@
-package com.comeon.backend.voting.command.domain;
-
-import java.time.LocalDate;
-
-public interface MeetingCalendarService {
-
-    boolean verifyDateInMeetingCalendar(Long meetingId, LocalDate date);
-}
