--- src/main/java/com/comeon/backend/voting/query/dto/VotingMemberSimple.java
@@ -1,16 +1,0 @@
-package com.comeon.backend.voting.query.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class VotingMemberSimple {
-
-    private Long userId;
-    private String nickname;
-    private String profileImageUrl;
-    private String memberRole;
-}
